+++ src/interpolation.lua
@@ -11,71 +11,190 @@
 ]]
 BeTweenApi.interpolation = {
 
-	-- straight increment from x to y in time.
+	--- straight increment from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
 	linear = function (x, y, t)
 		return x + t * (y - x)
 	end,
 
-	-- movement increment faster from x to y in time.
-	ease_in = function (x, y, t)
-		return x + ((t * t) / 1.0) * (y - x)
+	--- movement increment faster from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	quadratic_in = function (x, y, t)
+		return BeTweenApi.interpolation.linear(x, y, t ^ 2)
 	end,
 
-	-- movement increment slower from x to y in time.
-	ease_out = function (x, y, t)
-		t = 1 - ((1 - t) ^ 2)
+	--- movement increment slower from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	quadratic_out = function (x, y, t)
+		return BeTweenApi.interpolation.linear(x, y, t * (2.0 - t))
+	end,
+
+	--- movement is faster between x to y but slow down on begin and end in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	quadratic_in_out = function (x, y, t)
+		x = BeTweenApi.interpolation.quadratic_in(x, y, t)
+		y = BeTweenApi.interpolation.quadratic_out(x, y, t)
 		return BeTweenApi.interpolation.linear(x, y, t)
 	end,
 
-	-- movement is faster between x to y but slow down on begin and end in time.
-	ease_in_out = function (x, y, t)
-		x = BeTweenApi.interpolation.ease_in(x, y, t)
-		y = BeTweenApi.interpolation.ease_out(x, y, t)
-		return BeTweenApi.interpolation.linear(x, y, t)
+	--- movement increment faster from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	cubic_in = function (x, y, t)
+		return BeTweenApi.interpolation.linear(x, y, t ^ 3)
 	end,
 
-	-- movement jump from y to x and acting as a reverse elastic_out, movement is very slow but increment up in time.
-	elastic_in = function (x, y, t)
-		t = -(2.0 ^ (10.0 * (t - 1.0)) * math.sin((t - 0.1) * (2.0 * 3.14) / 0.4)) + 1.0
-		return BeTweenApi.interpolation.linear(x, y, t)
+	--- movement increment slower from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	cubic_out = function (x, y, t)
+		t = t - 1.0
+		return BeTweenApi.interpolation.linear(x, y, 1.0 + (t ^ 3))
 	end,
 
-	-- movement jump from x to y like an elastic, movement is very fast but slow down in time.
-	elastic_out = function (x, y, t)
+	--- movement is faster between x to y but slow down on begin and end in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	cubic_in_out = function (x, y, t)
+		
+		t = t * 2
+		if (t < 1.0) then
+			return BeTweenApi.interpolation.linear(x, y, 0.5 * (t ^ 3))
+		end
+
+		t = t - 2.0
+		return BeTweenApi.interpolation.linear(x, y, 0.5 * (t ^ 3 + 2.0))
+	end,
+
+	--- movement increment faster from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	quartic_in = function (x, y, t)
+		return BeTweenApi.interpolation.linear(x, y, t ^ 4)
+	end,
+
+	--- movement increment slower from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	quartic_out = function (x, y, t)
+		t = t - 1.0
+		return BeTweenApi.interpolation.linear(x, y, 1.0 - (t ^ 4))
+	end,
+
+	--- movement is faster between x to y but slow down on begin and end in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	quartic_in_out = function (x, y, t)
+		
+		t = t * 2
+		if (t < 1.0) then
+			return BeTweenApi.interpolation.linear(x, y, 0.5 * (t ^ 4))
+		end
+
+		t = t - 2.0
+		return BeTweenApi.interpolation.linear(x, y, -0.5 * (t ^ 4 - 2.0))
+	end,
+
+	--- movement jump from x to y like an elastic, movement is very fast but slow down in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	elastic = function (x, y, t)
 		t = (2.0 ^ -(10.0 * t) * math.sin((t - 0.1) * (2.0 * math.pi) / 0.4)) + 1.0
 		return BeTweenApi.interpolation.linear(x, y, t)
 	end,
 
-	-- movement moves like linear, but reach his destination in half the time, the rest of the time is used to come back to the starting point.
-	spike_linear = function (x, y, t)
-		if (t <= 0.5) then
-			return BeTweenApi.interpolation.linear(x, y, t / 0.5)
-		end
-		return BeTweenApi.interpolation.linear(x, y, (1 - t) / 0.5)
+	--- movement increment faster from x to y in time, the movement has the shape of a curve.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	sinusoidal_in = function (x, y, t)
+		t = 1.0 - math.cos(t * math.pi / 2.0)
+		return BeTweenApi.interpolation.linear(x, y, t)
 	end,
 
-	-- movement moves like ease in, but reach his destination in half the time, the rest of the time is used to come back to the starting point.
-	spike_ease_in = function (x, y, t)
-		if (t <= 0.5) then
-			return BeTweenApi.interpolation.linear(x, y, t / 0.5)
-		end
-		return BeTweenApi.interpolation.linear(x, y, (1 - t) / 0.5)
+	--- movement increment faster from x to y in time, the movement has the shape of a curve.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	sinusoidal_out = function (x, y, t)
+		t = math.sin(t * math.pi / 2.0)
+		return BeTweenApi.interpolation.linear(x, y, t)
 	end,
 
-	-- movement moves like ease out, but reach his destination in half the time, the rest of the time is used to come back to the starting point.
-	spike_ease_out = function (x, y, t)
-		if (t <= 0.5) then
-			return BeTweenApi.interpolation.ease_out(x, y, t / 0.5)
-		end
-		return BeTweenApi.interpolation.ease_out(x, y, (1 - t) / 0.5)
+	--- movement is faster between x to y but slow down on begin and end in time, the movement has the shape of a curve.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	sinusoidal_in_out = function (x, y, t)
+		t = 0.5 * (1.0 - math.cos(math.pi * t))
+		return BeTweenApi.interpolation.linear(x, y, t)
 	end,
 
-	-- movement moves like ease in out, but reach his destination in half the time, the rest of the time is used to come back to the starting point.
-	spike_ease_in_out = function (x, y, t)
-		if (t <= 0.5) then
-			return BeTweenApi.interpolation.ease_in_out(x, y, t / 0.5)
-		end
-		return BeTweenApi.interpolation.ease_in_out(x, y, (1 - t) / 0.5)
+
+	--- movement increment faster esponentially from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	circular_in  = function (x, y, t)
+		t = 1.0 - math.sqrt(1.0 - t * t)
+		return BeTweenApi.interpolation.linear(x, y, t)
 	end,
+
+	--- movement increment slower esponentially from x to y in time.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	circular_out = function (x, y, t)
+		t = t - 1.0
+		return BeTweenApi.interpolation.linear(x, y, math.sqrt(1.0 - (t * t)))
+	end,
+
+	--- movement increment slower from the start to the end but speed up esponentially in the middle.
+	--- @param x number
+	--- @param y number
+	--- @param t number
+	--- @return number
+	circular_in_out = function (x, y, t)
+
+		t = t * 2
+		if (t < 1.0) then
+			return BeTweenApi.interpolation.linear(x, y, -0.5 * (math.sqrt(1.0 - t * t) - 1.0))
+		end
+
+		t = t - 2.0
+		return BeTweenApi.interpolation.linear(x, y, 0.5 * (math.sqrt(1.0 - t * t) + 1.0))
+	end,
+	
 }